# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIError:
    properties:
      errorCode:
        type: integer
      errorMessage:
        type: string
    type: object
...
  schemas.Shipper:
    properties:
      companyName:
        example: Servientrega
        type: string
      phone:
        example: "6697901"
        type: string
    type: object
  schemas.ShipperResponse:
    properties:
      companyName:
        example: Servientrega
        type: string
      id:
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        type: string
      phone:
        example: "6697901"
        type: string
    type: object
  schemas.Subcategory:
    properties:
      categoryID:
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        type: string
      description:
        example: Description about subcategory 1
        type: string
      subCategoryName:
        example: Subcategory 1
        type: string
    type: object
  schemas.SubcategoryResponse:
    properties:
      categoryID:
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        type: string
      description:
        example: Description about subcategory 1
        type: string
      id:
        example: 2ld1f12f-2227-8s08-18cc-222fdb9634xx
        type: string
      subCategoryName:
        example: Subcategory 1
        type: string
    type: object
  schemas.Supplier:
    properties:
      address:
        example: 49 Spring St
        type: string
      city:
        example: Billings
        type: string
      companyName:
        example: Arthur Clark
        type: string
      country:
        example: United States
        type: string
      homePage:
        example: arthur.clark.com
        type: string
      phone:
        example: "6475240"
        type: string
      postalCode:
        example: "63104"
        type: string
    type: object
  schemas.SupplierResponse:
    properties:
      address:
        example: 49 Spring St
        type: string
      city:
        example: Billings
        type: string
      companyName:
        example: Arthur Clark
        type: string
      country:
        example: United States
        type: string
      homePage:
        example: arthur.clark.com
        type: string
      id:
        example: 2ld1f12f-2227-8s08-18cc-222fdb9634xx
        type: string
      phone:
        example: "6475240"
        type: string
      postalCode:
        example: "63104"
        type: string
    type: object
host: go-rest-api.onrender.com
info:
  contact:
    email: mario.diaz@correounivalle.edu.co
    name: AlexVonEinzbern
    url: https://github.com/AlexVonEinzbern
  description: ecommerce api-rest with basic CRUD operations (Create, Read, Update
    and Delete) products, suppliers, etc. Manage automatically invoice and shipping
    operations
  license:
    name: Apache 2.0
  termsOfService: http://swagger.io/terms/
  title: go-rest-api
  version: "1"
paths:
...
  /shippers:
    get:
      consumes:
      - application/json
      description: Search all shippers in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.ShipperResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all Shippers
      tags:
      - Shippers
    post:
      consumes:
      - application/json
      description: Create a Shipper
      parameters:
      - description: Shipper type
        in: body
        name: shipper
        required: true
        schema:
          $ref: '#/definitions/schemas.Shipper'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ShipperResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Shipper
      tags:
      - Shippers
  /shippers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a shipper by id
      parameters:
      - description: Shipper ID
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete Shipper by id
      tags:
      - Shippers
    patch:
      consumes:
      - application/json
      description: Update a shipper by id
      parameters:
      - description: Shipper ID
        example: 5dd1f36f-1627-4c88-98fb-601feb9634be
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/schemas.ShipperResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update Shipper by id
      tags:
      - Shippers
  /subcategories:
    get:
      consumes:
      - application/json
      description: Search all subcatagories in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.SubcategoryResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all Subcategories
      tags:
      - Subcategories
    post:
      consumes:
      - application/json
      description: Create a Subcategory
      parameters:
      - description: Create a subcategory
        in: body
        name: subcategory
        required: true
        schema:
          $ref: '#/definitions/schemas.Subcategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.SubcategoryResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Subcategory
      tags:
      - Subcategories
  /subcategories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a subcategory by id
      parameters:
      - description: Subcategory ID
        example: 2ld1f12f-2227-8s08-18cc-222fdb9634xx
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete a Subcategory by id
      tags:
      - Subcategories
    patch:
      consumes:
      - application/json
      description: Update a subcategory by id
      parameters:
      - description: Subcategory ID
        example: 2ld1f12f-2227-8s08-18cc-222fdb9634xx
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/schemas.SubcategoryResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update Subcategory by id
      tags:
      - Subcategories
  /subcategories/active:
    get:
      consumes:
      - application/json
      description: Search all active categories in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.SubcategoryResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all active Subcategories
      tags:
      - Subcategories
  /suppliers:
    get:
      consumes:
      - application/json
      description: Search all suppliers in the DataBase
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.SupplierResponse'
            type: array
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Search all Suppliers
      tags:
      - Suppliers
    post:
      consumes:
      - application/json
      description: Create a Supplier
      parameters:
      - description: Supplier type
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/schemas.Supplier'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.SupplierResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Create Supplier
      tags:
      - Suppliers
  /suppliers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a supplier by id
      parameters:
      - description: Supplier ID
        example: 2ld1f12f-2227-8s08-18cc-222fdb9634x
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete Supplier by id
      tags:
      - Suppliers
    patch:
      consumes:
      - application/json
      description: Update a supplier by id
      parameters:
      - description: Supplier ID
        example: 2ld1f12f-2227-8s08-18cc-222fdb9634x
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/schemas.SupplierResponse'
        "404":
          description: Can not find objects
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update Supplier by id
      tags:
      - Suppliers
swagger: "2.0"